% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdynfun.r
\name{addline}
\alias{addline}
\title{Function to add lines to landscape}
\usage{
addline(r, x1, x2, val, plot = FALSE)
}
\arguments{
\item{r}{raster that represents the landscape}

\item{x1}{from coordinates}

\item{x2}{to coordinates}

\item{val}{resistance value 1 is no resistance}

\item{plot}{switch if landscape should be plotted}
}
\value{
returns modified raster layer
}
\description{
adds a line from x1 to x2 to a raster layer
}
