% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdynfun.r
\name{addpoly}
\alias{addpoly}
\title{Function to add a polygon to a landscape}
\usage{
addpoly(r, pol, val, plot = T)
}
\arguments{
\item{r}{raster that represents the landscape}

\item{pol}{coordinates of the polygon}

\item{val}{resistance value, One equals no resistance}

\item{plot}{switch if landscape should be plotted}
}
\value{
returns modified raster layer
}
\description{
adds a polygon to a raster layer
}
