% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bilby}
\alias{bilby}
\title{Bilby data set}
\format{genlight object}
\usage{
bilby
}
\description{
This is a synthetic sample data set (in \code{\link{genind}} format) of
microsatellite data.
}
\author{
Bernd Gruber \email{bernd.gruber@canberra.edu.au}
}
\keyword{bilby}
\keyword{data(bilby)}
\keyword{datasets}
