% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popdynfun.r
\name{costdistances}
\alias{costdistances}
\title{Calculates cost distances for a given landscape (resistance matrix)}
\usage{
costdistances(landscape, locs, method, NN)
}
\arguments{
\item{landscape}{a raster object coding the resistance of the landscape}

\item{locs}{coordinates of the subpopulations}

\item{method}{defines the type of cost distance, types are "least-cost", "rSPDistance" or "commute (Circuitscape type)"}

\item{NN}{number of next neighbours recommendation is 8}
}
\value{
a costdistance matrix between all pairs of locs
}
\description{
calculates a cost distance matrix, to be used with run.popgensim
}

