% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popdynfun.r
\name{emigration}
\alias{emigration}
\title{Function to execute emigration on a pops object}
\usage{
emigration(xp, perc.mig, emi.m, emi.table = NULL)
}
\arguments{
\item{xp}{all pops combined in a list}

\item{perc.mig}{percentage if migrating individuals}

\item{emi.m}{emigration probability (normally based on cost dispersal distance)}

\item{emi.table}{a fixed number of migrating individuals can be specified (overrides emi.m)}
}
\value{
a list, first entry are updated pops, second entry the number of disperserin a matrix
}
\description{
emigration process on all population in one go
}

