\name{landgen}
\alias{landgen}
\docType{data}
\title{
A simulated genind data set with spatial coordinates 
}
\description{
This data set is used to demonstrate the use of the \code{\link{landgenreport}} and \code{\link{genleastcost}}  functions. It is a simple spatial \code{genind} object with 100 individuals in 10 populations and 20 loci with up to 20 alleles per loci.
}
\usage{data(landgen)}
\seealso{
 \code{\link{landgenreport}}, \code{\link{genleastcost}}
}
\examples{
data(landgen)
summary(landgen)
}
\keyword{datasets}
