% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popdynfun.r
\name{mutation}
\alias{mutation}
\title{Function to execute mutation on a pop data.frame}
\usage{
mutation(x, n.allels, mutrate, mtype = "n.all", n.cov)
}
\arguments{
\item{x}{a pop object}

\item{n.allels}{number of alleles}

\item{mutrate}{the mutation rate}

\item{mtype}{the mutation process (currently on n.all is implemented. (=Choose any allel from 1:n.allels))}

\item{n.cov}{(number of covariates, default  )}
}
\value{
a dispersal probability matrix
}
\description{
converts cost distances to probabilities: to reach a certain patch[ d0 average distance of p of all individuals, for example d0=100, p =0.5 -> 50\% procent of all
}

