% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdynfun.r
\name{p2p}
\alias{p2p}
\title{Function to calculate dispersal distances based on cost distances}
\usage{
p2p(x, d0, p)
}
\arguments{
\item{x}{Cost (Euclidean) distance matrix}

\item{d0}{dispersal distance}

\item{p}{of all individuals in a population}
}
\value{
a dispersal probability matrix
}
\description{
converts cost distances to probabilities: to reach a certain patch[ d0 average distance of p of all individuals, for example d0=100, p =0.5 -> 50\% procent of all migrating individuals go up to 100 m.
}

