% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdynfun.r
\name{pairwise.fstb}
\alias{pairwise.fstb}
\title{Calculates pairwise fsts using a genind object (very fast)}
\usage{
pairwise.fstb(gsp)
}
\arguments{
\item{gsp}{a genind object}
}
\value{
a pairwise fst matrix (same as hierfstat pairwise.fst)
}
\description{
for simulation the original pairwise.fst was too slow. The fast version works only with genind objects without NAs and diploid data (so to be save do not use it on non-simulated data)
}

