% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.propShared.r
\name{pairwise.propShared}
\alias{pairwise.propShared}
\title{Calculates proportion of shared alleles per pairs of populations}
\usage{
pairwise.propShared(gi)
}
\arguments{
\item{gi}{a genind object with at least two populations}
}
\value{
a matrix of proportion of shared alleles between populations
}
\description{
Calculates proportion of shared alleles per pairs of populations based on the minima of allele frequency for each allel (then summed and averaged over loci). Returns a similarity matrix (upper diagonal of the pairwise matrix.
}
