% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popdynfun.r
\name{pops2genind}
\alias{pops2genind}
\title{Function converts pops to a genind object}
\usage{
pops2genind(x, locs = NULL, n.cov = 3)
}
\arguments{
\item{x}{pops object (a list of pop)}

\item{locs}{named coordinates of locations}

\item{n.cov}{number of covariates (defaults to 3)}
}
\value{
a spatial genind object
}
\description{
converts pops into genind (to calculate Fst etc.)
}

