% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{possums}
\alias{possums}
\title{A genlight object created via the read.genetable functions [possum data set from Sarre et al. 2015]}
\format{genind object}
\usage{
possums
}
\description{
A genlight object created via the read.genetable functions [possum data set from Sarre et al. 2015]
}
\examples{
possums
}
\author{
Bernd Gruber \email{(bernd.gruber@canberra.edu.au}
}
\references{
Sarre, S.D., Aitken, N., Adamack, A.T., Macdonald, A.J., Gruber, B. & Cowan, P. (2014). Creating new evolutionary pathways through bioinvasion: The population genetics of brushtail possums in New Zealand. Molecular Ecology, 23, 3419-3433.
}
\keyword{datasets}

