% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdynfun.r
\name{reproduction}
\alias{reproduction}
\title{Function to execute reproduction on a pop data.frame}
\usage{
reproduction(x, type = "K.limit", K = n.ind, n.off, n.cov)
}
\arguments{
\item{x}{a pop object}

\item{type}{type of density dependence K.limit only yet}

\item{K}{Kapacity of a subpoplation}

\item{n.off}{number of offspring per female}

\item{n.cov}{(number of covariates, default 3 )}
}
\value{
an updated pop object
}
\description{
reproduction subprocess on single populations
}

