% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{treeScrub}
\alias{treeScrub}
\title{A raster file [proportion of treeScrub per cell in Hawke Bay, NZ]}
\format{data}
\usage{
treeScrub
}
\description{
A raster file [proportion of treeScrub per cell in Hawke Bay, NZ]
}
\references{
Sarre, S.D., Aitken, N., Adamack, A.T., Macdonald, A.J., Gruber, B. & Cowan, P. (2014). Creating new evolutionary pathways through bioinvasion: The population genetics of brushtail possums in New Zealand. Molecular Ecology, 23, 3419-3433.
}
\author{
Bernd Gruber \email{(bernd.gruber@canberra.edu.au}
}
\keyword{datasets}
