% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{fric.raster}
\alias{fric.raster}
\title{A simulated genind data set with spatial coordinates}
\format{An object of class \code{RasterLayer} of dimension 50 x 50 x 1.}
\usage{
fric.raster
}
\description{
This data set is used to demonstrate the use of the
\code{\link{landgenreport}} and \code{\link{genleastcost}} functions. It is
a simple spatial \code{genind} object with 100 individuals in 10 populations
and 20 loci with up to 20 alleles per loci.
}
\seealso{
\code{\link{landgenreport}}, \code{\link{genleastcost}}
}
\author{
Bernd Gruber \email{(bernd.gruber@canberra.edu.au}
}
\keyword{datasets}
