% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdynfun.r
\name{init.popgensim}
\alias{init.popgensim}
\title{Initialise a pops object fo a  a popgen simulation}
\usage{
init.popgensim(n.pops, n.ind, sex.ratio, n.loci, n.allels, locs, n.cov)
}
\arguments{
\item{n.pops}{number of subpopulations}

\item{n.ind}{number of individuals per subpopulation}

\item{sex.ratio}{sex ratio of males and females}

\item{n.loci}{number of loci}

\item{n.allels}{number of maximal alleles per loci}

\item{locs}{coordinates of the subpopulations}

\item{n.cov}{number of covariates (defaults to 3)}
}
\value{
a simpops object (to be used in run.popgensim )
}
\description{
initialises a time-forward, agent-based and spatiallly explicit genetic population simulation
}

