% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{landgen}
\alias{landgen}
\title{A simulated genind data set with spatial coordinates}
\format{An object of class \code{genind} of length 1.}
\usage{
landgen
}
\description{
This data set is used to demonstrate the use of the
\code{\link{landgenreport}} and \code{\link{genleastcost}} functions. It is
a simple spatial \code{genind} object with 100 individuals in 10 populations
and 20 loci with up to 20 alleles per loci.
}
\examples{
data(landgen)
summary(landgen)
}
\seealso{
\code{\link{landgenreport}}, \code{\link{genleastcost}}
}
\author{
Bernd Gruber \email{(bernd.gruber@canberra.edu.au}
}
\keyword{datasets}
