% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popdynfun.r
\name{pairwise.fstb}
\alias{pairwise.fstb}
\title{Calculates pairwise fsts using a genind object (very fast)}
\usage{
pairwise.fstb(gsp)
}
\arguments{
\item{gsp}{a genind object}
}
\value{
a pairwise fst matrix (same as hierfstat pairwise.fst)
}
\description{
for simulation the original pairwise.fst was much to slow. The fast version works only without NAs and diploid data (so do not use it on non-simulated data)
}

