\name{PopGen-package}
\alias{PopGen-package}
\alias{PopGen}
\docType{package}
\title{
Analysing population genetic data within a standardised framework
}     
\description{
This package aims to deliver a standardised framework for analysing population genetic data (at this stage only microsatellite data). The framework is open and meant to be user friendly, so R and its capabilities can be used by non-programmers. We encourage contributions and extensions of the package to allow for more sophisticated analyses of spatial populaiton genetic data. This packages draws heavily on already existing packages such as adegenet, mmod, pegas and others and we would like to thank the creators of those packages for their work in developing those packages.
}
\details{
\tabular{ll}{
Package: \tab PopGenReport\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2014-10-16\cr
License: \tab GPL\cr
}
The package provides a framework to analyse population genetic data. The main intention is to make the capabilities, which are already there, available to a wider community. Several standard routines are already implemented. We invite other developers and researchers to either recommend new extensions or analyses and to provide additional code that could be included within the existing framework which has a very modular format. The long term aim is to have a widely used standard package that serves as a first step in analyzing population genetic data and to facilitate comparability between studies.
}
\author{
Aaraon T. Adamack & Bernd Gruber (2012)

Maintainer: Bernd Gruber, Bernd.Gruber@canberra.edu.au

}
\references{
Adamack & Gruber (2014). PopGenReport - simplifying basic population genetic analyses in R, Methods in Ecology and Evolution,5, 384 - 387. doi: 10.1111/2041-210X.12158
}

\keyword{ package }
\seealso{
\code{\link{adegenet}}
\code{\link{mmod}}
\code{\link{pegas}}
}
